String Operations
Strings (MStrings) now supports concatenation using the + infix operator
this can be done with MStrings and any other object.
MStrings can also be compared to other objects, comparing agaist a different
type will only be equal if the MString is not empty and the other operand
is the Boolean true.
Also the lexer now validates literal MString input and returns invalid token
if the MString is interrupted by a line break or an unsupported scape sequence
is included.