Fixes #1 - Support for variable and constants
Now Monkey support the declaration of `let` (constants) and
`var` (variables). variables can be reassigned to new values
but trying to do so with `let` while cause an `AssignConstantError`.

Also once a variable or constant has been declared can't be
redeclared only assiged.

Finally assination will also type check meaning a variable with a
Integer value can't be assigned to a Boolean value..