Compile and execute functions as closures
We now have a new operation for pushing functions into the VM stack
this operation will treat all functions as clousures, currently closures
are not capturing variables but that is the next step.

Also to make this work the VM now provides a protocol to mark the base
type of the implementing language and default implementation of a
compiled function.